third_party/blink/web_tests/fast/css/remove-attribute-style-expected.txt

[bug 99295] removeAttribute('style') not working in certain circumstances. If this test passes, all backgroundColors are rgba(0, 0, 0, 0), i.e. all styles are removed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS elementWithoutStyleAttribute.style.backgroundColor = 'red'; elementWithoutStyleAttribute.removeAttribute('style'); getComputedStyle(elementWithoutStyleAttribute).backgroundColor is "rgba(0, 0, 0, 0)"
PASS getComputedStyle(elementWithEmptyStyleAttribute).backgroundColor is "rgba(0, 0, 0, 0)"
PASS getComputedStyle(elementWithStyleAttribute).backgroundColor is "rgba(0, 0, 0, 0)"
PASS elementWithoutStyleAttribute2.getAttribute('style') is null
PASS elementWithoutStyleAttribute2.style.backgroundColor = 'red'; elementWithoutStyleAttribute2.removeAttribute('style'); getComputedStyle(elementWithoutStyleAttribute2).backgroundColor is "rgba(0, 0, 0, 0)"
PASS elementWithoutStyleAttribute3.getAttribute('style') is "background-color: red;"
PASS getComputedStyle(elementWithoutStyleAttribute3).backgroundColor is "rgba(0, 0, 0, 0)"
PASS elementWithoutStyleAttribute4.style.backgroundColor = 'red'; elementWithoutStyleAttribute4.setAttribute('style', ''); elementWithoutStyleAttribute4.removeAttribute('style'); getComputedStyle(elementWithoutStyleAttribute4).backgroundColor is "rgba(0, 0, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE
